+++ config/services.php
@@ -27,7 +27,7 @@ return [
     'facebook' => [
         'client_id' => env('FACEBOOK_APP_ID'),
         'client_secret' => env('FACEBOOK_APP_SECRET'),
-        'redirect' => 'https://your-callback-url',
+        'redirect' => env('APP_URL') . 'callback/facebook',
     ],
 
     /*