+++ resources/views/auth/login.blade.php
@@ -9,7 +9,7 @@
             <div class="col-md-12">
                 <div class="breadcrumb-wrapper">
                     <a href="{{ URL::current() }}">
-                        <h1 class="product-title h2">Acceder</h1>
+                        <h1 class="product-title h2">Iniciar sesión en {{ config('app.name') }}</h1>
                     </a>
                 </div>
             </div>
@@ -18,12 +18,11 @@
 </div>
 <!-- Page Header End -->
 
-<!--
-    <ol class="breadcrumb">
-        <li><a href="index.html">Home /</a></li>
-        <li class="current">Login</li>
-    </ol>
--->
+<ol class="breadcrumb">
+    <li><a href="{{ config('app.url') }}">Inicio</a></li>
+    <li class="ml-2">/</li>
+    <li class="current ml-2">Iniciar sesión en {{ config('app.name') }}</li>
+</ol>
 
 @include('blocks.login')