+++ resources/views/blocks/login.blade.php
@@ -11,6 +11,16 @@
                     <div class="alert alert-danger">{{ session('message') }}</div>
                     @endif
 
+                    <div class="text-center m-3">
+                        <a class="btn btn-common btn-block text-white" href="{{ route('facebook_login') }}"><i class="lni-facebook-filled"></i> Acceder con Facebook</a>
+                    </div>
+
+                    <div class="text-center m-3">
+                        <a class="btn btn-common btn-block text-white" style="background-color: #1bd0ee" href="{{ route('twitter_login') }}"><i class="lni-twitter-filled"></i> Acceder con Twitter</a>
+                    </div>
+
+                    <hr>
+
                     <form role="form" class="login-form" method="POST" action="{{ route('login') }}">
 
                         @csrf